requirements.txt
tweepy-4.14.0